Updated XDEMO to latest version
Removed XDEMO 1.01 and updated it to 2.01.  The 2.x series doesn't use a
build script to create the accounts, all of the files are pre-populated.
The build script version caused problems with VSG on UniData.

There is one known issue.  In the MEMBERS dictionary change the
ENC_PARAMS record to ENC.PARAMS.  The DECRYPT.EXAMPLE program in U2BP
uses ENC.PARAMS.  This issue will be fixed in a future version.